Documenting, polishing

+++ count_sums.hpp
@@ -1,15 +1,15 @@
 
 #pragma once
 
-/// Calculates number of all combination with given sum
+/// Calculates number of all combination gives sum
 /// Input values treated as unique
 ///
 /// \param c values
 /// \param n sum
 /// \return sum combinations
-long sums_count(const std::vector<long>& c, int n)
+int sums_count(const std::vector<int>& c, int n)
 {
-    std::vector<long> numways(n + 1); // numways[x] means # ways to get sum x
+    std::vector<int> numways(n + 1); // numways[x] means # ways to get sum x
     numways[0] = 1; // init base case n=0
 
     // go thru coins 1-by-1 to build up numways[] dynamically
